backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.65.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "xlrd>=2.0.2",
    "prophet>=1.1.6",
    "groq>=1.0.0",
]
